src/fabric_jumpstart/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fabric-jumpstart"
authors = [{ name = "Microsoft Corporation" }]
version = "0.1.1"
description = "Get started quickly with pre-built solutions, tutorials, demos, and accelerators—automated, high-quality, and open-source."
readme = "README.md"
requires-python = ">=3.10,<=3.13"
license = "MIT"
urls.Repository = "https://github.com/microsoft/fabric-jumpstart.git"
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "fabric-cicd==0.1.34", # for testing local build: "fabric-cicd @ file:///C:/.../fabric-cicd/dist/fabric_cicd-0.1.34-py3-none-any.whl",
    "ipython>=8.20.0",
    "pyyaml>=6.0.3",
]

[tool.hatch.build]
include = ["fabric_jumpstart/**/*.py", "fabric_jumpstart/jumpstarts/**/*", "fabric_jumpstart/ui/**/*"]

[tool.hatch.build.targets.wheel]
packages = ["fabric_jumpstart"]

[tool.hatch.metadata]
allow-direct-references = true


[dependency-groups]
dev = [
    {include-group = "lint"},
    {include-group = "test"},
]
lint = [
    "ruff>=0.14.10",
    "ty==0.0.14",
]
test = [
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
]

[tool.ruff]
src = ["."]
# Ignore packaged jumpstart payloads (e.g., demo notebooks) to keep lint focused on library code.
extend-exclude = ["fabric_jumpstart/jumpstarts/**"]

[tool.ty.environment]
root = ["."]

[tool.ty.src]
# Skip type checking for test helpers that rely on dynamic imports/aliases.
exclude = [
    "tests/test_instance_name.py",
    "fabric_jumpstart/jumpstarts/**",
]
